Add interface between controller and service layer in listing
